move ini file creation to model
